BrowseProposal Page finished
I've finished the page for BrowseProposal(Professor) I'm also displaying the expired proposal.
I'm displaying only the logged in professor proposal and I've made each proposal  (row of the table cliccable), when the professor click on a row also the button modify is enable, but I haven't assigned any action to the modify button because we are not Implementing the update proposal story. I've modified the API in the backend to show only the proposal of the logged in  professor