Add code to auto-archive and send notifications to students when app is rejected
Now every minute a procedure is ran by psql that checks for expired thesis_proposals and sends a notification to the server with a list of all the application IDs that got rejected as a consequence of the archival of the newly expired thesis proposals.
The server listens for these notifications and proceed to send an email to all the students whose application got rejected.